# Added additional parameter to speed up processing. http://ghostscript…