….com/doc/current/Use.htm#Improving_performance

This makes a roughly 5x improvement in speed.

I tested in a batch script outside of islandora on a 1800 paged PDF.

*Without it:*
4607.11s user 166.27s system 99% cpu 1:19:59.61 total

*With the additional parameters:*
834.98s user 88.21s system 96% cpu 15:59.09 total

Also less errors where generated by the script.